Improve API for gradients over logistic circuits
Now instead of hacking around it, there are proper functions to call
Also, learned that a new cache is needed if evaluating the circuit for multiple inputs. Makes sense in hindsight.
Could altenatively evaluate it vectorized, but that messes with the gradients, will have to look into that one later